Move log to tencent_logger module
